combine actual_network with actual_network_data for iso7
