--- networks/iso7/actual_network_data.H
@@ -1,90 +1,0 @@
-#ifndef _actual_network_data_H_
-#define _actual_network_data_H_
-
-#include <AMReX_REAL.H>
-#include <AMREX_GpuArray.H>
-#include <fundamental_constants.H>
-#include <network_properties.H>
-
-using namespace amrex;
-using namespace Species;
-
-namespace iso7
-{
-    extern AMREX_GPU_MANAGED amrex::Array1D<amrex::Real, 1, NumSpec> bion;
-    extern AMREX_GPU_MANAGED amrex::Array1D<amrex::Real, 1, NumSpec> mion;
-    extern AMREX_GPU_MANAGED amrex::Array1D<amrex::Real, 1, NumSpec> wion;
-}
-
-namespace Rates {
-    enum NetworkRates {ircag = 1,
-                       iroga,
-                       ir3a,
-                       irg3a,
-                       ir1212,
-                       ir1216,
-                       ir1616,
-                       iroag,
-                       irnega,
-                       irneag,
-                       irmgga,
-                       irmgag,
-                       irsiga,
-                       ircaag,
-                       irtiga,
-                       irsi2ni,
-                       irni2si,
-                       NumRates = irni2si};
-
-    const int NGroups = 4;
-
-    extern amrex::Vector<std::string> names;
-};
-
-void actual_network_data_init()
-{
-    // Set the binding energy of the element
-    bion(He4)  = 28.29603e0_rt;
-    bion(C12)  = 92.16294e0_rt;
-    bion(O16)  = 127.62093e0_rt;
-    bion(Ne20) = 160.64788e0_rt;
-    bion(Mg24) = 198.25790e0_rt;
-    bion(Si28) = 236.53790e0_rt;
-    bion(Ni56) = 484.00300e0_rt;
-
-    // Set the mass
-    for (int i = 1; i <= NumSpec; ++i) {
-        mion(i) = nion(i) * C::m_n + zion[i-1] * (C::m_p + C::m_e) - bion(i) * C::MeV2gr;
-
-        // Molar mass
-        wion(i) = C::n_A * mion(i);
-
-        // Common approximation to molar mass
-        wion(i) = aion[i-1];
-    }
-
-    // set the names of the reaction rates
-    {
-        using namespace Rates;
-        names.resize(NumRates);
-        names(ircag)   = "rcag";
-        names(iroga)   = "roga";
-        names(ir3a)    = "r3a";
-        names(irg3a)   = "rg3a";    // inverse rate
-        names(ir1212)  = "r1212";
-        names(ir1216)  = "r1216";
-        names(ir1616)  = "r1616";
-        names(iroag)   = "roag";
-        names(irnega)  = "rnega";
-        names(irneag)  = "rneag";
-        names(irmgga)  = "rmgga";
-        names(irmgag)  = "rmgag";
-        names(irsiga)  = "rsiga";
-        names(ircaag)  = "rcaag";
-        names(irtiga)  = "rtiga";
-        names(irsi2ni) = "rsi2ni";
-        names(irni2si) = "rni2si";
-    }
-}
-
-#endif
